Local hardening is completly broken
The tutorial is 2 different one merged "Local hardening setup" and "Harden your project" are to replace "Installation", "Hardening"  and "Project setup" . IDK what to do with "Factory Testing". Once these works "Expand Your Buildtools" should be updated and finished.
Currently the build process does not work, i might be updated in the future, tt09 is broken and so please try these step again in the future.

There is left notes in lab5 about this.